Split recording area component into multiple files
